api/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
Flask==2.0.3
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.28
mysql-connector-python-rf==2.2.2
mysqlclient>=2.1.0
protobuf==3.19.4
PyMySQL==1.0.2
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.3
